refactor(connector): 重构元数据类型定义并优化文件集发现逻辑
重构接口中的元数据类型定义，将表、索引、外键等元数据拆分为独立类型
优化文件集发现逻辑，添加文件集列元数据支持
调整文件集资源处理逻辑，完善schema定义生成
