+++ adp/vega/vega-backend/server/logics/connectors/local/fileset/anyshare/anyshare.go
@@ -7,7 +7,6 @@
 package anyshare
 
 import (
-	"bytes"
 	"context"
 	"crypto/tls"
 	"encoding/json"
@@ -359,266 +358,4 @@ func (c *AnyShareConnector) GetMetadata(ctx context.Context) (map[string]any, er
 		"protocol":         c.config.Protocol,
 		"paths_configured": len(c.config.Paths) > 0,
 	}, nil
-}
-
-// buildFilesetMeta 创建 FilesetMeta 对象的辅助函数
-func buildFilesetMeta(id, name, displayPath, itemType, createdAt, modifiedAt string, createdBy, modifiedBy userInfoDTO, rev string) *interfaces.FilesetMeta {
-	return &interfaces.FilesetMeta{
-		ID:          id,
-		Name:        name,
-		DisplayPath: displayPath,
-		SourceMetadata: map[string]any{
-			"id":          id,
-			"name":        name,
-			"type":        itemType,
-			"created_at":  createdAt,
-			"modified_at": modifiedAt,
-			"created_by":  createdBy,
-			"modified_by": modifiedBy,
-			"rev":         rev,
-		},
-	}
-}
-
-// ListFilesets discovers one level of files/folders per configured roots (see design doc).
-func (c *AnyShareConnector) ListFilesets(ctx context.Context) ([]*interfaces.FilesetMeta, error) {
-	if err := c.Connect(ctx); err != nil {
-		return nil, err
-	}
-	var out []*interfaces.FilesetMeta
-
-	if len(c.config.Paths) == 0 {
-		// 获取知识库列表（仅第一层）
-		libs, err := c.getEntryDocLib(ctx)
-		if err != nil {
-			return nil, err
-		}
-
-		// 将知识库信息转换为FilesetMeta返回
-		for _, lib := range libs {
-			meta := buildFilesetMeta(lib.ID, lib.Name, lib.Name, lib.Type, lib.CreatedAt, lib.ModifiedAt, lib.CreatedBy, lib.ModifiedBy, lib.Rev)
-			out = append(out, meta)
-		}
-		return out, nil
-	}
-
-	// 遍历每个路径
-	for _, p := range c.config.Paths {
-
-		// 根据路径获取doc_id
-		docInfo, err := c.getDocIDByPath(ctx, p)
-		if err != nil {
-			return nil, err
-		}
-
-		// 检查路径是否指向文件（size != -1 表示文件）
-		if docInfo.Size != -1 {
-			return nil, fmt.Errorf("path %q must be a directory, not a file", p)
-		}
-
-		// 获取文件夹详细信息
-		detail, err := c.getDocItemDetail(ctx, docInfo.DocID)
-		if err != nil {
-			return nil, err
-		}
-
-		// 将文件夹本身作为一个 resource
-		meta := buildFilesetMeta(detail.DocID, detail.Name, detail.Path, detail.Type, detail.CreatedAt, detail.ModifiedAt, detail.CreatedBy, detail.ModifiedBy, detail.Rev)
-		out = append(out, meta)
-	}
-
-	return out, nil
-}
-
-// SearchFiles searches files based on query parameters.
-func (c *AnyShareConnector) SearchFiles(ctx context.Context, docID, keyword string, rows, start int) ([]map[string]any, int64, error) {
-	if err := c.Connect(ctx); err != nil {
-		return nil, 0, err
-	}
-	if docID == "" {
-		return nil, 0, fmt.Errorf("empty doc id")
-	}
-	u := fmt.Sprintf("%s/api/ecosearch/v1/file-search", c.baseURL)
-	payload := map[string]interface{}{
-		"keyword":      keyword,
-		"model":        "phrase",
-		"range":        []string{fmt.Sprintf("%s/*", docID)},
-		"rows":         rows,
-		"start":        start,
-		"type":         "doc",
-		"quick_search": true,
-	}
-	body, err := json.Marshal(payload)
-	if err != nil {
-		return nil, 0, err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, u, bytes.NewReader(body))
-	if err != nil {
-		return nil, 0, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", c.authHeader)
-
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return nil, 0, err
-	}
-	defer resp.Body.Close()
-	raw, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, 0, err
-	}
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return nil, 0, fmt.Errorf("file-search http %d: %s", resp.StatusCode, truncateForLog(raw))
-	}
-
-	var result struct {
-		Files []map[string]any `json:"files"`
-		Hits  int64            `json:"hits"`
-	}
-	if err := json.Unmarshal(raw, &result); err != nil {
-		return nil, 0, fmt.Errorf("file-search decode: %w", err)
-	}
-
-	return result.Files, result.Hits, nil
-}
-
-// GetObjectDownloadInfo returns download info for a given file or folder.
-func (c *AnyShareConnector) GetObjectDownloadInfo(ctx context.Context, resourceName, docID string) (map[string]any, error) {
-	if err := c.Connect(ctx); err != nil {
-		return nil, err
-	}
-	if docID == "" {
-		return nil, fmt.Errorf("empty doc id")
-	}
-	u := fmt.Sprintf("%s/api/open-doc/v1/file-download", c.baseURL)
-	// 当前 docID 为 gns://... 的形式，name为下载文件时的打包名称
-	payload := map[string]interface{}{
-		"doc": []map[string]string{
-			{"id": docID},
-		},
-		"name": resourceName,
-	}
-	body, err := json.Marshal(payload)
-	if err != nil {
-		return nil, err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, u, bytes.NewReader(body))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", c.authHeader)
-
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	raw, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return nil, fmt.Errorf("file-download http %d: %s", resp.StatusCode, truncateForLog(raw))
-	}
-	var m map[string]any
-	if err := json.Unmarshal(raw, &m); err != nil {
-		return nil, fmt.Errorf("file-download decode: %w", err)
-	}
-	if msg, ok := m["message"].(string); ok && msg != "" {
-		return nil, fmt.Errorf("file-download: %s", msg)
-	}
-	return m, nil
-}
-
-func (c *AnyShareConnector) getEntryDocLib(ctx context.Context) ([]docLibDTO, error) {
-	q := url.Values{}
-	q.Set("sort", "doc_lib_name")
-	q.Set("direction", "asc")
-	if c.config.DocLibType == docLibTypeKnowledge {
-		q.Set("type", "knowledge_doc_lib")
-	}
-	u := fmt.Sprintf("%s/api/document/v1/entry-doc-lib?%s", c.baseURL, q.Encode())
-	var libs []docLibDTO
-	if err := c.getJSON(ctx, u, &libs); err != nil {
-		return nil, err
-	}
-	return libs, nil
-}
-
-func (c *AnyShareConnector) getDocIDByPath(ctx context.Context, namepath string) (pathInfoDTO, error) {
-	u := fmt.Sprintf("%s/api/efast/v1/file/getinfobypath", c.baseURL)
-	body, err := json.Marshal(map[string]string{"namepath": namepath})
-
-	var info pathInfoDTO
-	if err != nil {
-		return info, err
-	}
-	req, err := http.NewRequestWithContext(ctx, http.MethodPost, u, bytes.NewReader(body))
-	if err != nil {
-		return info, err
-	}
-	req.Header.Set("Content-Type", "application/json")
-	req.Header.Set("Authorization", c.authHeader)
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return info, err
-	}
-	defer resp.Body.Close()
-	raw, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return info, err
-	}
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return info, fmt.Errorf("getinfobypath http %d: %s", resp.StatusCode, truncateForLog(raw))
-	}
-	if err := json.Unmarshal(raw, &info); err != nil {
-		return info, err
-	}
-	if info.DocID == "" {
-		return info, fmt.Errorf("getinfobypath: empty docid for path %q", namepath)
-	}
-	return info, nil
-}
-
-func (c *AnyShareConnector) getDocItemDetail(ctx context.Context, docID string) (*docItemDetailDTO, error) {
-	// docID格式为 gns://.../.../object_id，需要提取object_id
-	parts := strings.Split(docID, "/")
-	if len(parts) == 0 {
-		return nil, fmt.Errorf("invalid docID format: %s", docID)
-	}
-	objectID := parts[len(parts)-1]
-
-	u := fmt.Sprintf("%s/api/efast/v2/items/%s/all", c.baseURL, objectID)
-
-	var detail docItemDetailDTO
-	if err := c.getJSON(ctx, u, &detail); err != nil {
-		return nil, err
-	}
-	return &detail, nil
-}
-
-func (c *AnyShareConnector) getJSON(ctx context.Context, reqURL string, out any) error {
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, reqURL, nil)
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", c.authHeader)
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-	raw, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return err
-	}
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		return fmt.Errorf("GET %s: http %d %s", reqURL, resp.StatusCode, truncateForLog(raw))
-	}
-	if err := json.Unmarshal(raw, out); err != nil {
-		return fmt.Errorf("decode response: %w", err)
-	}
-	return nil
 }
