+++ adp/vega/vega-backend/server/logics/connectors/local/table/postgresql/postgresql_discover.go
@@ -220,7 +220,7 @@ ORDER BY ordinal_position`
 		return err
 	}
 
-	var columns []interfaces.ColumnMeta
+	var columns []interfaces.TableColumnMeta
 	for rows.Next() {
 		var name, dataType, udtName, isNullable sql.NullString
 		var colDefault, collation sql.NullString
@@ -243,7 +243,7 @@ ORDER BY ordinal_position`
 			orig = udtName.String
 		}
 
-		columns = append(columns, interfaces.ColumnMeta{
+		columns = append(columns, interfaces.TableColumnMeta{
 			Name:              name.String,
 			Type:              MapType(udtName.String, dataType.String),
 			OrigType:          orig,
@@ -322,7 +322,7 @@ ORDER BY i.relname, k.n`
 	}
 	defer rows.Close()
 
-	indexMap := make(map[string]*interfaces.IndexInfo)
+	indexMap := make(map[string]*interfaces.TableIndexMeta)
 	for rows.Next() {
 		var indexName, columnName string
 		var unique, primary bool
@@ -333,7 +333,7 @@ ORDER BY i.relname, k.n`
 		if idx, ok := indexMap[indexName]; ok {
 			idx.Columns = append(idx.Columns, columnName)
 		} else {
-			indexMap[indexName] = &interfaces.IndexInfo{
+			indexMap[indexName] = &interfaces.TableIndexMeta{
 				Name:    indexName,
 				Columns: []string{columnName},
 				Unique:  unique,
@@ -342,7 +342,7 @@ ORDER BY i.relname, k.n`
 		}
 	}
 
-	var indices []interfaces.IndexInfo
+	var indices []interfaces.TableIndexMeta
 	for _, idx := range indexMap {
 		indices = append(indices, *idx)
 	}
@@ -375,7 +375,7 @@ ORDER BY c.conname, u1.ord1`
 	}
 	defer rows.Close()
 
-	fkMap := make(map[string]*interfaces.ForeignKeyInfo)
+	fkMap := make(map[string]*interfaces.TableForeignKeyMeta)
 	for rows.Next() {
 		var cname, col, refCol, refSchema, refTable string
 		if err := rows.Scan(&cname, &col, &refCol, &refSchema, &refTable); err != nil {
@@ -386,7 +386,7 @@ ORDER BY c.conname, u1.ord1`
 			fk.Columns = append(fk.Columns, col)
 			fk.RefColumns = append(fk.RefColumns, refCol)
 		} else {
-			fkMap[cname] = &interfaces.ForeignKeyInfo{
+			fkMap[cname] = &interfaces.TableForeignKeyMeta{
 				Name:       cname,
 				Columns:    []string{col},
 				RefTable:   refFull,
@@ -395,7 +395,7 @@ ORDER BY c.conname, u1.ord1`
 		}
 	}
 
-	var fks []interfaces.ForeignKeyInfo
+	var fks []interfaces.TableForeignKeyMeta
 	for _, fk := range fkMap {
 		fks = append(fks, *fk)
 	}
